refactor: split flowrs-config lib.rs into auth, server, and paths modules
