+++ crates/flowrs-airflow-model/src/model/common/dagrun.rs
@@ -125,4 +125,3 @@ impl TimeBounded for DagRun {
         self.state == DagRunState::Running || self.state == DagRunState::Queued
     }
 }
-
