+++ crates/flowrs-airflow-model/src/model/common/taskinstance.rs
@@ -124,4 +124,3 @@ impl TimeBounded for TaskInstance {
         )
     }
 }
-
