+++ crates/flowrs-airflow/src/client/v1/mod.rs
@@ -12,9 +12,9 @@ use reqwest::Method;
 use url::{form_urlencoded, Url};
 
 use super::base::BaseClient;
-use flowrs_config::AirflowVersion;
-use flowrs_airflow_model::traits::AirflowClient;
 use flowrs_airflow_model::model::common::OpenItem;
+use flowrs_airflow_model::traits::AirflowClient;
+use flowrs_config::AirflowVersion;
 
 /// API v1 client implementation (for Airflow v2, uses /api/v1 endpoint)
 #[derive(Debug)]
