+++ src/app/model/config.rs
@@ -5,11 +5,11 @@ use ratatui::layout::{Constraint, Rect};
 use ratatui::text::Line;
 use ratatui::widgets::{Block, BorderType, Borders, Row, StatefulWidget, Table, Widget};
 
-use flowrs_config::AirflowConfig;
 use crate::app::events::custom::FlowrsEvent;
-use flowrs_airflow_model::model::common::OpenItem;
 use crate::app::worker::WorkerMessage;
 use crate::ui::theme::{BORDER_STYLE, SELECTED_ROW_STYLE, TABLE_HEADER_STYLE};
+use flowrs_airflow_model::model::common::OpenItem;
+use flowrs_config::AirflowConfig;
 
 use super::popup::config::commands::CONFIG_COMMAND_POP_UP;
 use super::{FilterableTable, KeyResult, Model, Popup};
