+++ src/app/model/dags.rs
@@ -9,7 +9,6 @@ use ratatui::text::{Line, Span};
 use ratatui::widgets::{Block, BorderType, Borders, Row, StatefulWidget, Table, Widget};
 use time::OffsetDateTime;
 
-use flowrs_airflow_model::model::common::{Dag, DagId, DagRunState, DagStatistic};
 use crate::app::events::custom::FlowrsEvent;
 use crate::app::model::popup::dagruns::trigger::TriggerDagRunPopUp;
 use crate::app::model::popup::dags::commands::DAG_COMMAND_POP_UP;
@@ -18,12 +17,13 @@ use crate::ui::constants::AirflowStateColor;
 use crate::ui::theme::{
     BORDER_STYLE, DAG_ACTIVE, SELECTED_ROW_STYLE, TABLE_HEADER_STYLE, TEXT_PRIMARY,
 };
+use flowrs_airflow_model::model::common::{Dag, DagId, DagRunState, DagStatistic};
 
 use super::dagruns::DagCodeView;
 use super::popup::dags::DagPopUp;
 use super::{FilterableTable, KeyResult, Model, Popup};
-use flowrs_airflow_model::model::common::OpenItem;
 use crate::app::worker::WorkerMessage;
+use flowrs_airflow_model::model::common::OpenItem;
 
 /// Model for the DAG panel, managing the list of DAGs and their filtering.
 pub struct DagModel {
