+++ src/app/model/taskinstances.rs
@@ -9,23 +9,23 @@ use ratatui::layout::{Constraint, Rect};
 use ratatui::text::Line;
 use ratatui::widgets::{Block, BorderType, Borders, Row, StatefulWidget, Table, Widget};
 
-use flowrs_airflow_model::graph::{sort_task_instances, TaskGraph};
-use flowrs_airflow_model::model::common::{
-    calculate_duration, format_duration, DagId, DagRunId, GanttData, TaskId, TaskInstance,
-    TaskInstanceState,
-};
 use crate::app::events::custom::FlowrsEvent;
 use crate::ui::common::{create_headers, state_to_colored_square};
 use crate::ui::constants::AirflowStateColor;
 use crate::ui::gantt::create_gantt_bar;
 use crate::ui::theme::{BORDER_STYLE, SELECTED_ROW_STYLE, TABLE_HEADER_STYLE};
+use flowrs_airflow_model::graph::{sort_task_instances, TaskGraph};
+use flowrs_airflow_model::model::common::{
+    calculate_duration, format_duration, DagId, DagRunId, GanttData, TaskId, TaskInstance,
+    TaskInstanceState,
+};
 
 use super::popup::taskinstances::clear::ClearTaskInstancePopup;
 use super::popup::taskinstances::mark::MarkTaskInstancePopup;
 use super::popup::taskinstances::TaskInstancePopUp;
 use super::{FilterableTable, KeyResult, Model, Popup};
-use flowrs_airflow_model::model::common::OpenItem;
 use crate::app::worker::WorkerMessage;
+use flowrs_airflow_model::model::common::OpenItem;
 
 /// Model for the Task Instance panel, managing the list of task instances and their filtering.
 pub struct TaskInstanceModel {
