+++ src/app/worker/dagruns.rs
@@ -2,10 +2,10 @@ use std::sync::{Arc, Mutex};
 
 use log::debug;
 
-use flowrs_airflow_model::model::common::{DagId, DagRunId, DagRunState};
-use flowrs_airflow_model::traits::AirflowClient;
 use crate::app::model::popup::dagruns::mark::MarkState;
 use crate::app::state::App;
+use flowrs_airflow_model::model::common::{DagId, DagRunId, DagRunState};
+use flowrs_airflow_model::traits::AirflowClient;
 
 /// Handle updating the list of DAG runs for a specific DAG.
 ///
