+++ src/app/worker/mod.rs
@@ -4,8 +4,8 @@ use std::sync::{Arc, Mutex};
 use super::model::popup::dagruns::mark::MarkState;
 use super::model::popup::taskinstances::mark::MarkState as TaskMarkState;
 use super::state::App;
-use flowrs_airflow_model::model::common::{DagId, DagRunId, OpenItem, TaskId};
 use anyhow::Result;
+use flowrs_airflow_model::model::common::{DagId, DagRunId, OpenItem, TaskId};
 use tokio::sync::mpsc::Receiver;
 use tokio::task::JoinSet;
 
@@ -73,7 +73,6 @@ pub enum WorkerMessage {
     },
     OpenItem(OpenItem),
 }
-
 
 impl WorkerMessage {
     /// Returns a dedup key for periodic refresh messages.
