+++ src/app/worker/tasks.rs
@@ -2,9 +2,9 @@ use std::sync::{Arc, Mutex};
 
 use log::debug;
 
+use crate::app::state::App;
 use flowrs_airflow_model::graph::TaskGraph;
 use flowrs_airflow_model::traits::AirflowClient;
-use crate::app::state::App;
 
 /// Handle fetching task definitions and building the task graph
 pub async fn handle_update_tasks(
