+++ src/commands/run.rs
@@ -8,12 +8,12 @@ use crossterm::event::{DisableFocusChange, EnableFocusChange};
 use crossterm::ExecutableCommand;
 use log::{info, LevelFilter, Log, Metadata, Record};
 
-use flowrs_config::FlowrsConfig;
-use flowrs_airflow::managed_services::expand::expand_managed_services;
 use crate::app::run_app;
 use crate::app::state::App;
 use crate::CONFIG_PATHS;
 use anyhow::Result;
+use flowrs_airflow::managed_services::expand::expand_managed_services;
+use flowrs_config::FlowrsConfig;
 
 struct FileLogger {
     file: Mutex<File>,
